# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ursol CAST API - ERP System",
        "description": "API completa de ERP para empresas costarricenses. Sistema integral de gestión empresarial con soporte para contabilidad, facturación electrónica, inventario, compras, ventas, recursos humanos y más.",
        "contact": {
            "name": "Sistemas Ursol",
            "email": "soporte@ursol.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de Desarrollo Local"
        },
        {
            "url": "https://api.ursol.com",
            "description": "Servidor de Producción"
        }
    ],
    "paths": {
...
        "/api/empleados": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Listar empleados",
                "description": "Lista empleados de la empresa autenticada con filtros opcionales",
                "operationId": "8eb442fb48a9986f146dfe2d189e0f35",
                "parameters": [
                    {
                        "name": "activo",
                        "in": "query",
                        "description": "Filtrar por estado activo",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "cargo_id",
                        "in": "query",
                        "description": "Filtrar por cargo",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Empleado"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Crear empleado",
                "description": "Registra un nuevo empleado en la empresa autenticada",
                "operationId": "07a84cb11dd1dc24fce5f1b7abcd09e4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "apellidos",
                                    "numero_identificacion"
                                ],
                                "properties": {
                                    "cargo_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "numero_identificacion": {
                                        "type": "string",
                                        "example": "1-2345-6789"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Carlos"
                                    },
                                    "apellidos": {
                                        "type": "string",
                                        "example": "Rodríguez"
                                    },
                                    "email": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "fecha_ingreso": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "salario_base": {
                                        "type": "number",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Empleado creado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/empleados/{id}": {
            "get": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Obtener empleado",
                "description": "Detalles de un empleado",
                "operationId": "5bcf15ebe357d0f0539eb8ab7e13cdae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Empleado encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Empleado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Actualizar empleado",
                "description": "Actualizar un empleado existente\n\nPUT/PATCH /api/empleados/{id}",
                "operationId": "f7428f349f0166fdf8f46320ccd1f669",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "apellidos": {
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "salario_base": {
                                        "type": "number",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Actualizado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Empleados"
                ],
                "summary": "Eliminar empleado",
                "description": "Soft delete del empleado",
                "operationId": "866f9e07c5da23613f5a60f136deed36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Eliminado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cargo": {
                "title": "Cargo",
                "description": "Cargo o puesto de trabajo",
                "required": [
                    "id",
                    "nombre",
                    "empresa_id"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Gerente General"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Responsable de la administración general",
                        "nullable": true
                    },
                    "empresa_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
...
            "Empleado": {
                "title": "Empleado",
                "description": "Empleado de la empresa",
                "required": [
                    "id",
                    "empresa_id",
                    "nombre",
                    "apellidos",
                    "numero_identificacion"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "empresa_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cargo_id": {
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "numero_identificacion": {
                        "type": "string",
                        "example": "1-2345-6789"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Carlos"
                    },
                    "apellidos": {
                        "type": "string",
                        "example": "Rodríguez Mora"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "carlos.rodriguez@empresa.com",
                        "nullable": true
                    },
                    "telefono": {
                        "type": "string",
                        "example": "8888-9999",
                        "nullable": true
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Heredia, San Francisco",
                        "nullable": true
                    },
                    "fecha_ingreso": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-15",
                        "nullable": true
                    },
                    "salario_base": {
                        "type": "number",
                        "format": "decimal",
                        "example": 650000,
                        "nullable": true
                    },
                    "activo": {
                        "type": "boolean",
                        "example": true
                    },
                    "eliminado": {
                        "type": "boolean",
                        "example": false
                    },
                    "creado_en": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "actualizado_en": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "cargo": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/Cargo"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum Bearer Token Authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}